requirements.txt
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.39.0
openai==0.28.0
lxml==4.9.3
validators==0.20.0
tqdm==4.66.1